BugFix: Ignore hidden files in workspaces folder
All files saved under the workspaces directory are displayed in the open workspace
prompt. Under certain circumstances, unwanted files would be displayed accordingly.

On Mac OS X for instance, a .DS_STORE file can be created under the workspaces
directory which will obfuscate the true workspaces.

This commit ignores any 'hidden' folders within the workspaces directory. As an alternative
fix, we could have only shown files that were with the '.ws' extension, but this may be
too restrictive under some circumstances. Perhaps the user changes the extension in the settings?

We can revisit as needed if this solution proves insufficient.